TTT-20: Refactors 4 filenames.
