+++ tests/com/bl/ttt/test_game.py
@@ -1,7 +1,7 @@
 """
 Tests for the Game class
 """
-from com.bl.ttt.ttt_game import Game
+from com.bl.ttt.game import Game
 
 
 def test_01_is_any_row_complete_identifies_complete_row():
